requirements1.txt
appdirs==1.4.0
asn1crypto==0.22.0
awscli==1.11.102
boto3==1.4.4
botocore==1.5.80
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
colorama==0.3.7
cryptography==1.9
cycler==0.10.0
defusedxml==0.5.0
Django==1.10.5
django-allauth==0.32.0
django-filter==1.0.1
django-registration==2.2
django-registration-redux==1.6
django-rest-auth==0.9.1
djangorestframework==3.5.3
docutils==0.13.1
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
gunicorn==19.6.0
idna==2.5
ipaddress==1.0.18
jmespath==0.9.3
Markdown==2.6.7
matplotlib==2.0.2
MySQL-python==1.2.5
mysqlclient==1.3.10
numpy==1.13.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pandas==0.20.3
Pillow==4.0.0
pip-review==0.5.3
pyasn1==0.2.3
pycparser==2.17
pyfcm==1.3.1
pyOpenSSL==17.0.0
pyparsing==2.1.10
python-dateutil==2.6.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.18.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rsa==3.4.2
s3transfer==0.1.10
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
sklearn==0.0
subprocess32==3.2.7
urllib3==1.21.1
